feat: Phase UI-1H — cleanup + migration to canonical URLs
- /my-characters-new renamed to /my-characters (gear_plan_pages.py)
- Old /my-characters route removed from profile_pages.py
- /gear-plan now returns HTTP 302 → /my-characters
- Gear Plan nav link removed from base.html
- my_characters_new.{html,css,js} renamed to my_characters.{html,css,js}
- Old my_characters.{html,css,js} and gear_plan.{html,css,js} deleted
- Migration 0081: retires my_gear_plan screen_permission, adds my_characters
- Unit tests updated for phases 5.0/5.1/5.2/5.4 to match new mcn-* classes
