+++ src/guild_portal/templates/base.html
@@ -26,7 +26,6 @@
                     <a href="/roster" class="nav-link">Roster</a>
                     <a href="/crafting-corner" class="nav-link">Crafting</a>
                     <a href="/my-characters" class="nav-link">My Characters</a>
-                    <a href="/gear-plan" class="nav-link">Gear Plan</a>
                     <a href="/profile" class="nav-link">Settings</a>
                     <span class="nav-badge nav-badge--char">{{ current_member.main_character.character_name if current_member.main_character else current_member.display_name }}</span>
                     {% if current_member.guild_rank %}
