--- src/guild_portal/templates/member/gear_plan.html
@@ -1,112 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Gear Plan — {{ site().guild_name }}{% endblock %}
-
-{% block head %}
-<link rel="stylesheet" href="/static/css/gear_plan.css?v=12">
-{% endblock %}
-
-{% block content %}
-<div class="gp-page">
-
-  {# ── Page header ─────────────────────────────────────────────── #}
-  <div class="page-title-row">
-    <h1 class="gp-page__title">Gear Plan</h1>
-    <div class="page-title-actions">
-      <select id="gp-char-select" class="gp-select gp-select--char" hidden></select>
-    </div>
-  </div>
-
-  {# ── Status / error ───────────────────────────────────────────── #}
-  <div id="gp-loading" class="gp-state-msg">Loading characters…</div>
-  <div id="gp-error"   class="gp-state-msg gp-state-msg--error" hidden></div>
-  <div id="gp-no-chars" class="gp-state-msg" hidden>
-    You have no claimed characters yet.
-    <a href="/profile" class="btn btn-primary" style="margin-left:1rem">Go to Settings → Character Claims</a>
-  </div>
-
-  {# ── Main paperdoll layout ────────────────────────────────────── #}
-  <div id="gp-main" class="gp-main" hidden>
-
-    {# Left column: Head → Wrist (shirt + tabard as inactive cosmetic) #}
-    <div class="gp-col gp-col--left" id="gp-col-left"></div>
-
-    {# Centre: char summary + controls + weapons at bottom #}
-    <div class="gp-center" id="gp-center">
-      <div class="gp-char-badge" id="gp-char-badge">
-        <span id="gp-char-name" class="gp-char-badge__name"></span>
-        <span id="gp-char-meta" class="gp-char-badge__meta"></span>
-      </div>
-
-      <div class="gp-controls">
-        <div class="gp-ctrl-row">
-          <label class="gp-label">Hero Talent</label>
-          <select id="gp-ht-select" class="gp-select">
-            <option value="">— Any —</option>
-          </select>
-        </div>
-        <div class="gp-ctrl-row">
-          <label class="gp-label">BIS Source</label>
-          <select id="gp-source-select" class="gp-select"></select>
-        </div>
-        <div class="gp-ctrl-actions">
-          <button id="gp-btn-populate"     class="btn btn-primary btn-sm">Fill BIS</button>
-          <button id="gp-btn-sync"         class="btn btn-secondary btn-sm">Sync Gear</button>
-          <button id="gp-btn-import-simc"  class="btn btn-secondary btn-sm">Import SimC</button>
-          <button id="gp-btn-export-simc"  class="btn btn-secondary btn-sm">Export SimC</button>
-          <button id="gp-btn-delete-plan"  class="btn btn-danger btn-sm">Reset Plan</button>
-        </div>
-      </div>
-
-      <div id="gp-status" class="gp-status" hidden></div>
-
-      {# Weapons row — main hand + off hand, anchored to grid bottom #}
-      <div id="gp-col-weapons" class="gp-weapons-row"></div>
-    </div>
-
-    {# Right column: Hands → Trinket 2 #}
-    <div class="gp-col gp-col--right" id="gp-col-right"></div>
-
-  </div>
-
-  {# ── Slot detail drawer ───────────────────────────────────────── #}
-  <div id="gp-drawer" class="gp-drawer" hidden>
-    <div class="gp-drawer__header">
-      <span id="gp-drawer-title" class="gp-drawer__title"></span>
-      <button id="gp-drawer-close" class="gp-drawer__close">&times;</button>
-    </div>
-    <div id="gp-drawer-body" class="gp-drawer__body"></div>
-  </div>
-
-</div>
-
-{# ── SimC Import Modal ────────────────────────────────────────── #}
-<div id="gp-simc-modal" class="gp-modal" hidden>
-  <div class="gp-modal__backdrop"></div>
-  <div class="gp-modal__box">
-    <div class="gp-modal__header">
-      <h3 class="gp-modal__title">Import SimC Profile</h3>
-      <button id="gp-simc-close" class="gp-modal__close">&times;</button>
-    </div>
-    <div class="gp-modal__body">
-      <p class="gp-modal__hint">
-        Paste your SimC profile from the in-game SimulationCraft addon,
-        <a href="https://archon.gg" target="_blank" rel="noopener">Archon</a>, or
-        <a href="https://www.wowhead.com" target="_blank" rel="noopener">Wowhead</a>.
-        Locked slots will not be overwritten.
-      </p>
-      <textarea id="gp-simc-text" class="gp-simc-textarea"
-                placeholder="Paste SimC profile here…" rows="12"></textarea>
-    </div>
-    <div class="gp-modal__footer">
-      <button id="gp-simc-submit" class="btn btn-primary">Import</button>
-      <button id="gp-simc-cancel" class="btn btn-secondary">Cancel</button>
-    </div>
-  </div>
-</div>
-{% endblock %}
-
-{% block scripts %}
-<script>var whTooltips = {colorLinks: true, iconizeLinks: false, renameLinks: false, hide: {sellprice: true, ilvl: false, reqlevel: false, power: false}};</script>
-<script src="https://wow.zamimg.com/widgets/power.js"></script>
-<script src="/static/js/gear_plan.js?v=12"></script>
-{% endblock %}
